+++ include/platform.h
@@ -17,16 +17,16 @@
 class Platform : public Interactable
 {
 public:
-    Platform(const int tileWidth);
+    Platform(const uint32_t tileWidth);
     Platform(std::string save); // Loading from save
     ~Platform();
 
     void operator=(const Platform&) = delete;
 
-    bool canPlace(const Vect<int>& pos, std::vector<std::unique_ptr<Interactable>>& objects, const Vect<int>& size) override;
+    bool canPlace(const Vect<int64_t>& pos, std::vector<std::unique_ptr<Interactable>>& objects, const Vect<uint32_t>& size) override;
 
     // void update(const uint64_t& seconds) override;
-    void render(Window& window, const Vect<int>& renderOffset) override;
+    void render(Window& window, const Vect<int64_t>& renderOffset) override;
 
     std::string getSave() override;
     std::string readSave(std::string& save) override;
